+++ customtkinter/widgets/ctk_frame.py
@@ -3,7 +3,7 @@ from typing import Union, Tuple, List
 from .core_rendering.ctk_canvas import CTkCanvas
 from ..theme_manager import ThemeManager
 from .core_rendering.draw_engine import DrawEngine
-from .widget_base_class import CTkBaseClass
+from .core_widget_classes.widget_base_class import CTkBaseClass
 
 
 class CTkFrame(CTkBaseClass):