+++ customtkinter/widgets/ctk_optionmenu.py
@@ -5,9 +5,9 @@ from typing import Union, Tuple, Callable
 from .core_rendering.ctk_canvas import CTkCanvas
 from ..theme_manager import ThemeManager
 from .core_rendering.draw_engine import DrawEngine
-from .widget_base_class import CTkBaseClass
-from .dropdown_menu import DropdownMenu
-from ..utility.ctk_font import CTkFont
+from .core_widget_classes.widget_base_class import CTkBaseClass
+from .core_widget_classes.dropdown_menu import DropdownMenu
+from .font.ctk_font import CTkFont
 
 
 class CTkOptionMenu(CTkBaseClass):