+++ customtkinter/widgets/ctk_progressbar.py
@@ -5,7 +5,7 @@ from typing import Union, Tuple
 from .core_rendering.ctk_canvas import CTkCanvas
 from ..theme_manager import ThemeManager
 from .core_rendering.draw_engine import DrawEngine
-from .widget_base_class import CTkBaseClass
+from .core_widget_classes.widget_base_class import CTkBaseClass
 
 
 class CTkProgressBar(CTkBaseClass):