+++ customtkinter/widgets/ctk_tabview.py
@@ -2,7 +2,7 @@ from typing import Union, Tuple, Dict, List, Callable
 
 from ..theme_manager import ThemeManager
 from .ctk_frame import CTkFrame
-from .widget_base_class import CTkBaseClass
+from .core_widget_classes.widget_base_class import CTkBaseClass
 from .ctk_segmented_button import CTkSegmentedButton
 from .core_rendering.ctk_canvas import CTkCanvas
 from .core_rendering.draw_engine import DrawEngine