+++ customtkinter/widgets/ctk_textbox.py
@@ -5,8 +5,8 @@ from .core_rendering.ctk_canvas import CTkCanvas
 from .ctk_scrollbar import CTkScrollbar
 from ..theme_manager import ThemeManager
 from .core_rendering.draw_engine import DrawEngine
-from .widget_base_class import CTkBaseClass
-from ..utility.ctk_font import CTkFont
+from .core_widget_classes.widget_base_class import CTkBaseClass
+from .font.ctk_font import CTkFont
 
 from customtkinter.utility.utility_functions import pop_from_dict_by_set, check_kwargs_empty